requirements.txt
apache-beam==2.5.0
dateinfer==0.2.0
google-cloud==0.33.1
google-cloud-dataflow==2.5.0
google-cloud-bigquery==1.4.0
phonenumbers==8.9.10
firebase-admin==2.8.0
gcsfs==0.1.1
six==1.10.0
gapic-google-cloud-pubsub-v1==0.15.4
gax-google-pubsub-v1==0.8.3
google-cloud-pubsub==0.35.4
google-gax==0.15.15
googleapis-common-protos==1.5.3
grpc-google-pubsub-v1==0.8.1
grpcio==1.13.0
proto-google-cloud-pubsub-v1==0.15.4
protobuf==3.6.0